+++ example/lib/main.dart
@@ -76,7 +76,7 @@ final class CounterLoaded extends CounterState {
   CounterLoaded({required this.count});
 }
 
-final class CounterBlocError extends CounterState implements UnexpectedErrorAPI {
+final class CounterBlocError extends CounterState implements UnexpectedErrorBase {
   @override
   final UnexpectedError error;
 
@@ -92,8 +92,11 @@ class CounterBloc extends SafeBloc<CounterEvent, CounterState> {
     onSafe<IncrementPressed>(_increment, isAction: true);
   }
 
-  FutureOr<void> _increment(IncrementPressed event, SafeEmitter<CounterState> emit,
-      {required String trackingId}) async {
+  FutureOr<void> _increment(
+    IncrementPressed event,
+    SafeEmitter<CounterState> emit, {
+    required String trackingId,
+  }) async {
     if (state case final CounterLoaded loaded) {
       final incrementedCount = loaded.count + 1;