+++ lib/src/unexpected_error_handler.dart
@@ -19,16 +19,16 @@ typedef UnexpectedErrorActionCallback = Future<void> Function(
   UnexpectedError error,
 );
 
-/// Widget that listens to the events and states that implement [UnexpectedErrorAPI] in specified the [BLOC].
+/// Widget that listens to the events and states that implement [UnexpectedErrorBase] in specified the [BLOC].
 /// Subsequently, it calls [onErrorAction] in case of error event
 /// or displays a [errorScreen] in case of error state.
-class UnexpectedErrorHandler<BLOC extends SafeBlocBase<STATE>, STATE> extends StatelessWidget {
+class UnexpectedErrorHandler<BLOC extends SafeBlocBase<STATE, BaseEffect>, STATE> extends StatelessWidget {
   final Widget child;
 
-  /// Callback that returns a [Widget] that is displayed if state that implements [UnexpectedErrorAPI] is emitted by [BLOC].
+  /// Callback that returns a [Widget] that is displayed if state that implements [UnexpectedErrorBase] is emitted by [BLOC].
   final UnexpectedErrorWidgetBuilder? errorScreen;
 
-  /// Callback that is invoked if event that implements [UnexpectedErrorAPI] is emitted by [BLOC].
+  /// Callback that is invoked if event that implements [UnexpectedErrorBase] is emitted by [BLOC].
   final UnexpectedErrorActionCallback? onErrorAction;
 
   const UnexpectedErrorHandler({
@@ -39,9 +39,9 @@ class UnexpectedErrorHandler<BLOC extends SafeBlocBase<STATE>, STATE> extends St
   });
 
   void _handleError(BuildContext context, BaseEffect effect) {
-    // BaseEffect can implement UnexpectedErrorAPI
+    // BaseEffect can implement UnexpectedErrorBase
     // ignore: avoid-unrelated-type-assertions
-    if (effect case final UnexpectedErrorAPI error) {
+    if (effect case final UnexpectedErrorBase error) {
       if (onErrorAction != null) {
         unawaited(onErrorAction?.call(context, error.error));
       } else {
@@ -85,7 +85,7 @@ class UnexpectedErrorHandler<BLOC extends SafeBlocBase<STATE>, STATE> extends St
       listener: _handleError,
       child: BlocBuilder<BLOC, STATE>(
         builder: (context, state) {
-          if (state is UnexpectedErrorAPI && !state.error.isAction) {
+          if (state is UnexpectedErrorBase && !state.error.isAction) {
             return Material(
               child: Builder(
                 builder: (context) {