+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 1.0.0-rc.2
+- `UnexpectedErrorAPI` renamed to -> `UnexpectedErrorBase`
+- bloc presentation test added for testing blocs and cubits with presentation
+- `SafeBlocWithPresentation` and `SafeCubitWithPresentation` added
+
 ## 1.0.0-rc.1
 - Initialize package.